+++ package.json
@@ -22,7 +22,7 @@
 		"oxfmt": "^0.26.0",
 		"semver": "^7.5.0",
 		"tshy": "^3.0.2",
-		"tsup": "^6.7.0",
+		"tsdown": "^0.21.4",
 		"tsx": "^4.7.0",
 		"typescript": "^5.8.3",
 		"vite": "^5.0.2",
