+++ packages/hub/tsdown.config.ts
@@ -1,22 +1,23 @@
-import type { Options } from "tsup";
+import type { Options } from "tsdown";
 
-const baseConfig: Options = {
+const baseConfig = {
 	entry: ["./index.ts"],
 	format: ["cjs", "esm"],
 	outDir: "dist",
 	clean: true,
-};
+} satisfies Options;
 
 const nodeConfig: Options = {
 	...baseConfig,
+	entry: [...baseConfig.entry, "./cli.ts"],
 	platform: "node",
 };
 
 const browserConfig: Options = {
 	...baseConfig,
 	platform: "browser",
-	target: "es2018",
-	splitting: true,
+	target: "es2022",
+
 	outDir: "dist/browser",
 };
