+++ packages/ollama-utils/package.json
@@ -40,7 +40,7 @@
 		"format": "oxfmt .",
 		"format:check": "oxfmt --check .",
 		"prepublishOnly": "pnpm run build",
-		"build": "tsup src/index.ts --format cjs,esm --clean && tsc --emitDeclarationOnly --declaration",
+		"build": "tsdown src/index.ts --format cjs,esm --clean && tsc --emitDeclarationOnly --declaration",
 		"build:automap": "tsx scripts/generate-automap.ts && oxfmt ./src/chat-template-automap.ts",
 		"test": "vitest run",
 		"check": "tsc"
