+++ packages/space-header/package.json
@@ -34,7 +34,7 @@
 		"access": "public"
 	},
 	"scripts": {
-		"build": "tsup && tsc --emitDeclarationOnly --declaration",
+		"build": "tsdown && tsc --emitDeclarationOnly --declaration",
 		"lint": "eslint --quiet --fix --ext .cjs,.ts .",
 		"lint:check": "eslint --ext .cjs,.ts .",
 		"format": "oxfmt .",
@@ -42,7 +42,7 @@
 		"prepublishOnly": "pnpm run build"
 	},
 	"devDependencies": {
-		"tsup": "^8.1.0",
+		"tsdown": "^0.21.4",
 		"typescript": "^5.4.5"
 	},
 	"packageManager": "pnpm@10.10.0"
