+++ packages/space-header/tsdown.config.ts
@@ -1,24 +1,21 @@
-import type { Options } from "tsup";
+import type { UserConfig } from "tsdown";
 
-const baseConfig: Options = {
-	entry: ["./index.ts"],
+const baseConfig: UserConfig = {
+	entry: ["./src/index.ts"],
 	format: ["cjs", "esm"],
 	outDir: "dist",
 	clean: true,
 };
 
-const nodeConfig: Options = {
+const nodeConfig: UserConfig = {
 	...baseConfig,
-	entry: ["./index.ts"],
 	platform: "node",
 };
 
-const browserConfig: Options = {
+const browserConfig: UserConfig = {
 	...baseConfig,
-	entry: ["./index.ts"],
 	platform: "browser",
 	target: "es2018",
-	splitting: true,
 	outDir: "dist/browser",
 };
