+++ packages/space-header/tsup.config.d.ts
@@ -1,3 +1,3 @@
-import type { Options } from "tsup";
+import type { Options } from "tsdown";
 declare const _default: Options[];
 export default _default;
